Make focus follow when moving desk to another display.
(Presumably?) Previously it was neccessary to manually move the focus after
moving a desktop to another display (like you have to do when moving a window).
That appears nolonger to be the case and now when you move a desk the focus
automatically follows. This commit removes focus changing from the keybinding
which recently has been irritatingly switching the focus to the wrong
display...